Update codebase to use nextjs i18n
Closes #1984
