+++ frontend/src/pages/[slug]/index.tsx
