+++ frontend/src/pages/[slug]/page.graphql
