+++ frontend/src/pages/blog/[slug]/blog-post.graphql
