+++ frontend/src/pages/blog/[slug]/blog-social-card.graphql
