+++ frontend/src/pages/cfp/is-cfp-open.graphql
