+++ frontend/src/pages/cfp/me-submissions.graphql
