+++ frontend/src/pages/login/social-login-check.graphql
