+++ frontend/src/pages/reset-password/[id]/[token].tsx
