+++ frontend/src/pages/schedule/[day].tsx
