+++ frontend/src/pages/schedule/add-schedule-slot.graphql
