+++ frontend/src/pages/schedule/fragments/schedule-slot.graphql
