+++ frontend/src/pages/schedule/update-or-create-item.graphql
