+++ frontend/src/pages/submission/[id]/index.tsx
