+++ frontend/src/pages/submission/[id]/is-voting-closed.graphql
