+++ frontend/src/pages/talk/[slug]/talk-social-card.graphql
