+++ frontend/src/pages/tickets/information.tsx
