+++ frontend/src/pages/voting/index.tsx
