+++ frontend/src/pages/voting/send-vote.graphql
