+++ frontend/src/pages/voting/voting-submissions.graphql
